refactor: decompose hooks into multiple files
