--- bricks/r13n/hooks/.gitignore
@@ -1,4 +1,0 @@
-.dart_tool
-.packages
-pubspec.lock
-build/
