+++ bricks/r13n/hooks/r13n_exception.dart
@@ -1,0 +1,8 @@
+abstract class R13nException implements Exception {
+  const R13nException({required this.message});
+
+  final String message;
+
+  @override
+  String toString() => '$runtimeType: $message';
+}
