+++ bricks/r13n/hooks/test/src/arb_value_test.dart
@@ -2,6 +2,7 @@
 
 import 'package:test/test.dart';
 
+import '../../hooks.dart';
 import '../../pre_gen.dart';
 
 void main() {
