+++ bricks/r13n/hooks/test/src/exceptions/r13n_yaml_not_found_exception_test.dart
@@ -1,0 +1,21 @@
+import 'package:test/test.dart';
+
+import '../../../hooks.dart';
+
+void main() {
+  group('R13nYamlNotFoundException', () {
+    test('can be instantiated', () {
+      final error = R13nYamlNotFoundException();
+      expect(error, isA<R13nYamlNotFoundException>());
+      expect(error.message, equals('No r13n.yaml found.'));
+    });
+
+    test('toString() ', () {
+      final error = R13nYamlNotFoundException();
+      expect(
+        error.toString(),
+        equals('R13nYamlNotFoundException: No r13n.yaml found.'),
+      );
+    });
+  });
+}
