--- bricks/r13n/hooks/test/src/exceptions/yaml_not_found_exception_test.dart
@@ -1,21 +1,0 @@
-import 'package:test/test.dart';
-
-import '../../../pre_gen.dart';
-
-void main() {
-  group('YamlNotFoundException', () {
-    test('can be instantiated', () {
-      final error = YamlNotFoundException();
-      expect(error, isA<YamlNotFoundException>());
-      expect(error.message, equals('No r13n.yaml found.'));
-    });
-
-    test('toString() ', () {
-      final error = YamlNotFoundException();
-      expect(
-        error.toString(),
-        equals('YamlNotFoundException: No r13n.yaml found.'),
-      );
-    });
-  });
-}
